clean up pyflask design
