--- pyflask/apis/__init__.py
@@ -1,2 +1,0 @@
-from .apiStartup import api as startup_resource
-from .apiNeuroConv import api as neuroconv_resource
