--- pyflask/errorHandlers/__init__.py
@@ -1,1 +1,0 @@
-from .notBadRequestException import notBadRequestException
