--- pyflask/errorHandlers/notBadRequestException.py
@@ -1,5 +1,0 @@
-def notBadRequestException(exception):
-    """
-    Check if the exception is a generic exception.
-    """
-    return type(exception).__name__ not in ["BadRequest", "Forbidden", "Unauthorized"]
