--- pyflask/namespaces/__init__.py
@@ -1,2 +1,0 @@
-from .configureNamespaces import configure_namespaces, get_namespace, NamespaceEnum
-from .namespaceLogger import get_namespace_logger
