--- pyflask/setupUtils/__init__.py
@@ -1,3 +1,0 @@
-from .configureAPI import configureAPI
-from .configureLogger import configureLogger
-from .configureRouteHandlers import configureRouteHandlers
