--- pyflask/startup/__init__.py
@@ -1,1 +1,0 @@
-from .minimumApiVersion import get_api_version
