--- pyflask/startup/minimumApiVersion.py
@@ -1,9 +1,0 @@
-import os
-
-
-def get_api_version():
-    """
-    Returns the version of the API
-    """
-
-    return {"version": os.getenv("API_VERSION", "9.4.1")}
