+++ pyflask/tools/__init__.py
@@ -1,4 +1,4 @@
-from .manage_neuroconv import (
+from .neuroconv_helpers import (
     get_all_interface_info,
     locate_data,
     get_source_schema,
